backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
social-auth-app-django==5.4.0
